+++ imports/ui/components/Task.tsx
@@ -4,11 +4,10 @@ import {customEvent} from '/imports/ui/model/formIo'
 import {Tasks} from '/imports/api/tasks'
 import {global_Error, Meteor} from 'meteor/meteor'
 import {useTracker} from 'meteor/react-meteor-data'
-import {Alert, Button, Loader} from "@epfl/epfl-sti-react-library"
-import {Link} from "react-router-dom"
+import {Button, Loader} from "@epfl/epfl-sti-react-library"
+import {Link, useHistory} from "react-router-dom"
 import _ from "lodash"
 import {findDisabledFields} from "/imports/lib/formIOUtils";
-import { useHistory } from "react-router-dom";
 import toast from 'react-hot-toast';
 import {ErrorIcon} from "react-hot-toast/src/components/error";
 import {toastClosable} from "/imports/ui/components/Toasters";
@@ -59,7 +58,7 @@ export function Task({workflowKey}: { workflowKey: string }) {
     return (<>
       {!toBeSubmitted ? (
         <>
-          <Alert message={'Data submitted !'} alertType={'success'}/>
+          <div className={'alert alert-success'} role='alert'>{'Data submitted !'}</div>
           <Link to={`/`}><Button label={'Back'} onClickFn={() => void 0}/></Link>
         </>
       ) : (